requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pymongo==4.6.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart
streamlit